refactor by extracting the http request method functions
